HW1 - KN
I had a lot of problems pushing the file but Chelsea was a lot of help. I also greatly appreciated the Markdown tutorial. After forking the file I saved the file and opened it in R studio. I then made my edits and pushed it to the repository with Git desktop.