Update README.md
This assignment was not too difficult. I found the commands of markdown to be very intuitive, save the linking of an image. I am not used to finding the file path of an image on Mac, so I will have to get better at that. I also lack intuitive understanding of the interaction between what I do in RStudio and what is on GitHub. Thus, when I wrote the code I kept trying to run it, so as to make sure I got the assignment right. I had to read more carefully to see that I would have to see the output on GitHub. This led to a mistake that I then had to fix (removing an image that didn't link properly).