Saniya
For me, this exercise was very challenging. 
At first when I created my Github account, there was a tutorial for using Github that included a similar exercise of editing the readme file with our introduction. I got confused if this was the same thing as the assignment or not. 
Then I went back to instructions and saw that we had to create a *locally* to Readme in Rstudio. 
There was a mention of using Markdown so I downloaded Github for Desktop as well as Atom and starting typing my commands following the syntax provided in *Mastering Markdown*. However, I could see no "pull" requests in either Atom or Github so I realised I was doing something wrong. 
Then I went back to the README file and opened it in Rstudio this time, made my changes and saved it. I couldn't understand how to commit my changes to my repo. Whenever I went to the homework page and tried making a *Pull request*, I was not able to. I kept using Google to answer my problems and referring to the different instruction manuals uploaded.
I went back to Github for Desktop and saw that my changes were appearing there and I made a push request there. Then I came back to *fork* and saw my edited version uploaded here.
I am still not able to upload the image from my desktop to this.

I think reading different instruction manuals at the same time created more confusion so I should be very mindful of establishing the order of steps before working from next time.